+++ src/encoded/static/components/item-pages/CaseView/InterpretationTab.js
@@ -1,12 +1,7 @@
 'use strict';
 
 import React, { useCallback, useMemo, useState } from 'react';
-import queryString from 'query-string';
-import moment from 'moment';
 import DropdownButton from 'react-bootstrap/esm/DropdownButton';
-import { LocalizedTime } from '@hms-dbmi-bgm/shared-portal-components/es/components/ui/LocalizedTime';
-import { Checkbox } from '@hms-dbmi-bgm/shared-portal-components/es/components/forms/components/Checkbox';
-import { variantSampleColumnExtensionMap, VariantSampleDisplayTitleColumn } from './../../browse/variantSampleColumnExtensionMap';
 import { VariantSampleSelection, VariantSampleSelectionList } from './VariantSampleSelection';
